Movement fixed, need to fix input and rotation. Details in description
Movement was heavily bugged out due to influencing player rotation in CameraController(?!). All code relating to that has been commented out, proper friction value applied (around 6 seems to be solid) and movement is functional. Need to look into the inputs as the left and right input seems to be reversed!