all Runes assets added!
all the runes are there divided by their element! pls ignore the one that sais runes, go for the one that sais prefab!(the othe other one has texture issues but are the original ones so they cant be deleted otherwise the good ones will be gone! ill fix it!)